feat: restructure file upload
