+++ stories/FilesUpload.stories.tsx
@@ -1,7 +1,7 @@
 import { faker } from "@faker-js/faker";
 import { ActionFunctionArgs } from "react-router";
 import type { Meta, StoryObj } from "@storybook/react";
-import FilesUpload from "~/components/filesUpload/FilesUpload";
+import FilesUpload from "~/components/formElements/filesUpload/FilesUpload";
 import {
   PDFFileMetadata,
   TEN_MB_IN_BYTES,
